Final Version as of presentation
Note, the database included is connected via a local direct address. The connection string must be changed if this program is to be run on a different program. This is due to the way the code executes from the debug folder and will make a copy of the database. Using a reference address causes the database readonly.